mongodb1/requirements.txt
pymongo>=4.0
pandas>=1.3
pyyaml>=5.4
